+++ e2e/tests/plugins.e2e.js
@@ -81,10 +81,20 @@ describe('plugins', () => {
       );
 
       it('expand gallery image on full screen', function() {
-        cy.get(`[data-hook=${'image-item'}]:last`)
+        cy.get(`[data-hook=${'image-item'}]`)
+          .eq(2)
           .parent()
           .click();
+        cy.get(
+          '#pgi65a6266ba23a8a55da3f469157f15237_0 > div > div > div > a > div > canvas'
+        ).should('be.visible');
         cy.eyesCheckWindow({ tag: this.test.title, target: 'window', fully: false });
+        cy.get(`[data-hook=${'nav-arrow-next'}]`).click({ force: true });
+        cy.get(
+          '#pgiea8ec1609e052b7f196935318316299d_1 > div > div > div > a > div > canvas'
+        ).should('be.visible');
+        cy.get(`[data-hook=${'fullscreen-close-button'}]`).click();
+        cy.eyesCheckWindow({ tag: 'closed fullscreen', target: 'window', fully: false });
       });
     });
   });
@@ -210,7 +220,9 @@ describe('plugins', () => {
       cy.openVideoUploadModal().addVideoFromURL();
       cy.shrinkPlugin();
       cy.waitForVideoToLoad();
-      cy.focusEditor().enterParagraphs(['Will this fix the flakiness?']);
+      cy.getEditor()
+        .type('{uparrow}')
+        .type('Will this fix the flakiness?');
       cy.eyesCheckWindow(this.test.title);
     });
 
@@ -218,7 +230,9 @@ describe('plugins', () => {
       cy.openVideoUploadModal().addCustomVideo();
       cy.shrinkPlugin();
       cy.waitForVideoToLoad();
-      cy.focusEditor().enterParagraphs(['Will this fix the flakiness?']);
+      cy.getEditor()
+        .type('{uparrow}')
+        .type('Will this fix the flakiness?');
 
       cy.eyesCheckWindow(this.test.title);
     });
@@ -243,9 +257,11 @@ describe('plugins', () => {
 
     it('add a soundcloud URL', function() {
       cy.openSoundCloudModal().addSoundCloud();
-      cy.waitForVideoToLoad();
       cy.shrinkPlugin();
-      cy.focusEditor().enterParagraphs(['Will this fix the flakiness?']);
+      cy.waitForVideoToLoad();
+      cy.getEditor()
+        .type('{uparrow}')
+        .type('Will this fix the flakiness?');
       cy.eyesCheckWindow(this.test.title);
     });
   });