+++ packages/editor/web/src/RichContentEditor/I18nRichContentEditor.jsx
@@ -18,5 +18,3 @@ export default class I18nRichContentEditor extends Component {
     return <WrappedEditor {...this.props} ref={this.setEditorRef} />;
   }
 }
-
-I18nRichContentEditor.displayName = 'RichContentEditor';