+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -95,7 +95,7 @@ class RichContentEditor extends Component {
       locale,
       anchorTarget,
       relValue,
-      helpers,
+      helpers = {},
       config,
       isMobile = false,
       shouldRenderOptimizedImages,