+++ packages/plugin-gallery/web/src/createGalleryPlugin.js
@@ -3,10 +3,12 @@ import { createBasePlugin } from 'wix-rich-content-editor-common';
 import { Component, DEFAULTS } from './gallery-component';
 import { GALLERY_TYPE } from './types';
 
+const fileInputAccept = 'image/* || video/*';
+
 const createGalleryPlugin = (config = {}) => {
   const type = GALLERY_TYPE;
   const { helpers, theme, t, anchorTarget, relValue, [type]: settings = {}, ...rest } = config;
-
+  settings.accept = settings.accept || fileInputAccept;
   return createBasePlugin({
     component: Component,
     settings,