+++ packages/plugin-gallery/web/src/gallery-component.jsx
@@ -1,8 +1,9 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
+import { Loader } from 'wix-rich-content-editor-common';
 import { isEqual } from 'lodash';
 import GalleryViewer from './gallery-viewer';
-import { DEFAULTS } from './constants';
+import { DEFAULTS, imageItem } from './constants';
 
 //eslint-disable-next-line no-unused-vars
 const EMPTY_SMALL_PLACEHOLDER =
@@ -37,26 +38,29 @@ class GalleryComponent extends PureComponent {
       !isEqual(componentState, nextProps.componentState)
     ) {
       this.setState(this.stateFromProps(nextProps));
+    } else if (componentData.items?.length > 0) {
+      this.onLoad(false);
     }
   }
 
   stateFromProps = props => {
     const items = props.componentData.items || []; // || DEFAULTS.items;
     const styles = { ...DEFAULTS.styles, ...(props.componentData.styles || {}) };
-    const isLoading = (props.componentState && props.componentState.isLoading) || 0;
+    const itemsLeftToUpload = props.componentState?.isLoading || 0;
     const state = {
       items,
       styles,
-      isLoading,
+      itemsLeftToUpload,
     };
 
     if (props.componentState) {
       const { userSelectedFiles } = props.componentState;
-      if (isLoading <= 0 && userSelectedFiles) {
+      if (itemsLeftToUpload <= 0 && userSelectedFiles) {
         //lets continue the uploading process
         if (userSelectedFiles.files && userSelectedFiles.files.length > 0) {
-          state.isLoading = userSelectedFiles.files.length;
+          state.itemsLeftToUpload = userSelectedFiles.files.length;
           this.handleFilesSelected(userSelectedFiles.files);
+          state.isLoading = true;
         }
         if (this.props.store) {
           setTimeout(() => {
@@ -105,19 +109,12 @@ class GalleryComponent extends PureComponent {
       reader.onload = e => this.fileLoaded(e, file, itemPos);
       reader.readAsDataURL(file);
     });
+    this.state && this.onLoad(true);
   };
 
   imageLoaded = (event, file, itemPos) => {
     const img = event.target;
-    const item = {
-      metadata: {
-        height: img.height,
-        width: img.width,
-      },
-      itemId: String(event.timeStamp),
-      url: img.src,
-    };
-
+    const item = imageItem(img, String(event.timeStamp));
     const itemIdx = this.setItemInGallery(item, itemPos);
     const { helpers } = this.props;
     const hasFileChangeHelper = helpers && helpers.onFilesChange;
@@ -133,12 +130,16 @@ class GalleryComponent extends PureComponent {
     const handleFileAdded = (item, idx) => {
       const galleryItem = {
         metadata: {
+          type: item.type || 'image',
           height: item.height,
           width: item.width,
         },
         itemId: String(item.id),
         url: item.file_name,
       };
+      if (item.type === 'video') {
+        galleryItem.metadata.poster = item.poster || item.thumbnail_url;
+      }
       this.setItemInGallery(galleryItem, idx);
     };
 
@@ -151,27 +152,57 @@ class GalleryComponent extends PureComponent {
     }
   };
 
+  videoLoaded = (event, file, itemPos) => {
+    const { helpers } = this.context;
+    const hasFileChangeHelper = helpers && helpers.onVideoSelected;
+
+    if (hasFileChangeHelper) {
+      helpers.onVideoSelected(file, video => {
+        // eslint-disable-next-line camelcase
+        const data = { ...video, id: String(event.timeStamp), file_name: video.video_url };
+        this.handleFilesAdded({ data, itemPos });
+      });
+    } else {
+      console.warn('Missing upload function'); //eslint-disable-line no-console
+    }
+  };
+
   fileLoaded = (event, file, itemPos) => {
-    const img = new Image();
-    img.onload = e => this.imageLoaded(e, file, itemPos);
-    img.src = event.target.result;
+    if (file.type.match('image/*')) {
+      const img = new Image();
+      img.onload = e => this.imageLoaded(e, file, itemPos);
+      img.src = event.target.result;
+    } else if (file.type.match('video/*')) {
+      this.videoLoaded(event, file, itemPos);
+    }
+  };
+
+  renderLoader = () => {
+    return <Loader type={'medium'} />;
+  };
+
+  onLoad = isLoading => {
+    this.setState({ isLoading });
   };
 
   render() {
     return (
-      <GalleryViewer
-        componentData={this.props.componentData}
-        onClick={this.props.onClick}
-        className={this.props.className}
-        settings={this.props.settings}
-        theme={this.props.theme}
-        helpers={this.props.helpers}
-        disableRightClick={this.props.disableRightClick}
-        isMobile={this.props.isMobile}
-        anchorTarget={this.props.anchorTarget}
-        relValue={this.props.relValue}
-        blockKey={this.blockKey}
-      />
+      <>
+        <GalleryViewer
+          componentData={this.props.componentData}
+          onClick={this.props.onClick}
+          className={this.props.className}
+          settings={this.props.settings}
+          theme={this.props.theme}
+          helpers={this.props.helpers}
+          disableRightClick={this.props.disableRightClick}
+          isMobile={this.props.isMobile}
+          anchorTarget={this.props.anchorTarget}
+          relValue={this.props.relValue}
+          blockKey={this.blockKey}
+        />
+        {this.state.isLoading && this.renderLoader()}
+      </>
     );
   }
 }