+++ scripts/templates/atomic-plugin/web/src/typeMapper.js
@@ -1,0 +1,6 @@
+import YourPluginNameViewer from './yourPluginName-viewer';
+import { YOUR_PLUGIN_NAME_TYPE } from './types';
+
+export const typeMapper = () => ({
+  [YOUR_PLUGIN_NAME_TYPE]: { component: YourPluginNameViewer },
+});