+++ CHANGELOG.md
@@ -13,6 +13,18 @@
     Click to see more.
   </summary>
 
+
+</details>
+<hr/>
+
+## 6.10.0 (Mar 11, 2020)
+### :rocket: New Feature
+- `gallery`
+  - [#625](https://github.com/wix-incubator/rich-content/pull/625) gallery supports adding videos
+- `fullscreen`
+  - [#625](https://github.com/wix-incubator/rich-content/pull/625) supports target prop to render fullscreen at the target element
+- `editor-common`
+  - [#716](https://github.com/wix-incubator/rich-content/pull/716) progress loader
 ### :bug: Bug Fix
 - `hashtag`
   - [#707](https://github.com/wix-incubator/rich-content/pull/707) allow hashtag and link in the same block
@@ -25,8 +37,11 @@
   - [#757](https://github.com/wix-incubator/rich-content/pull/757) prevent underline link inline style to affects on new line
 - `link`
   - [#756](https://github.com/wix-incubator/rich-content/pull/756) selection keeps in place when canceling link panel
-</details>
-<hr/>
+- `gallery`
+  - [#697](https://github.com/wix-incubator/rich-content/pull/697) altText keeping value when changing to next image
+### :house: Internal
+- `fullscreen`
+  - [#625](https://github.com/wix-incubator/rich-content/pull/625) replaced react-images with Pro-Gallery
 
 ## 6.9.6 (Mar 5, 2020)
 ### :bug: Bug Fix